BUG/MINOR: Fix the sending function in Lua's cosocket
This is a regression from the commit a73e59b6901a164d19b1145e8511602d9814f28f.

When data are sent from a cosocket, the action is done in the context of the
applet running a lua stack and not in the context of the applet owning the
cosocket. So we must take care, explicitly, that this last applet have a buffer
(the req buffer of the cosocket).

This patch must be backported in 1.7
(cherry picked from commit 33834b15dce4658ee98fa85668162d78abf32e18)
(cherry picked from commit 8d570235f18eb7a6c920d674d1a057f35c6740e8)

Signed-off-by: Willy Tarreau <w@1wt.eu>