BUG/MINOR: stream: Don't forget to remove CF_WAKE_ONCE flag on response channel
This flag can be set on a channel to pretend there is activity on it. This is a
way to wake-up the corresponding stream and evaluate stream analyzers on the
channel. It is correctly handled on both channels but removed only on the
request channel.

This patch is flagged as a bug but for now, CF_WAKE_ONCE is never set on the
response channel.

(cherry picked from commit cdaea89a0c0eb8c028ce9220c1af7ea795845dd7)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>
(cherry picked from commit 7de6df3a6624ba5b023d76700fd1ee12a634dc6a)

Signed-off-by: Willy Tarreau <w@1wt.eu>