BUG/MEDIUM: tcp-check: don't call tcpcheck_main() from the I/O handlers!
This function can destroy a socket and create a new one, resulting in a
change of FD on the connection between recv() and send() for example,
which is absolutely not permitted, and can result in various funny games
like polling not being properly updated (or with the flags from a previous
fd) etc.

Let's only call this from the wake() callback which is more tolerant.
Ideally the operations should be made even more reliable by returning
a specific value to indicate that the connection was released and that
another one was created. But this is hasardous for stable releases as
it may reveal other issues.

This fix should be backported to 1.7 and 1.6.

(cherry picked from commit c09572fd8be6e7d7a1207bca31e14856ee7ac9d2)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>
(cherry picked from commit dbdfe96dd3090c009963b7119ae204c19fe8438c)
Signed-off-by: Willy Tarreau <w@1wt.eu>