BUG/MEDIUM: tcp-check: single connect rule can't detect DOWN servers
When tcpcheck is used to do TCP port monitoring only and the script is
composed by a single "tcp-check connect" rule (whatever port and ssl
options enabled), then the server can't be seen as DOWN.
Simple configuration to reproduce:

  backend b
    [...]
    option tcp-check
    tcp-check connect
    server s1 127.0.0.1:22 check

The main reason for this issue is that the piece of code which validates
that we're not at the end of the chained list (of rules) prevents
executing the validation of the establishment of the TCP connection.
Since validation is not executed, the rule is terminated and the report
says no errors were encountered, hence the server is UP all the time.

The workaround is simple: move the connection validation outsied the
CONNECT rule processing loop, into the main function.
That way, if the connection status is not CONNECTED, then HAProxy will
now add more time to wait for it. If the time is expired, an error is
now well reported.

Backport status: 1.8

(cherry picked from commit 248f1173f272b10bf7ed2aa7061bfafd34160f70)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 69ea5b29b63c3d891844e94b37d8cca94d04dedf)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit c76beb7403fd4b320dda8255cd482de924037d8c)
[wt: it in fact needed to be backported further]
Signed-off-by: Willy Tarreau <w@1wt.eu>