BUG/MAJOR: stream-int: Update the stream expiration date in stream_int_notify()
Since a long time, the expiration date of a stream is only updated in
process_stream(). It is calculated, among others, using the channels expiration
dates for reads and writes (.rex and .wex values). But these values are updated
by the stream-interface. So when this happens at the connection layer, the
update is only done if the stream's task is woken up. Otherwise, the stream
expiration date is not immediatly updated. This leads to unexpected
behaviours. Time to time, users reported that the wrong timeout was hitted or
the wrong termination state was reported. This is partly because of this
bug.

Recently, we observed some blocked sessions for a while when big objects are
served from the cache applet. It seems only concern the clients not reading the
response. Because delivered objects are big, not all data can be sent. And
because delivered objects are big, data are fast forwarded (from the input to
the output with no stream wakeup). So in such situation, the stream expiration
date is never updated and no timeout is hitted. The session remains blocked
while the client remains connected.

This bug exists at least since HAProxy 1.5. But recent changes on the connection
layer make it more visible. It must be backported from 1.9 to 1.6. And with more
pain it should be backported to 1.5.

(cherry picked from commit d7607de06574f00e999434051e082615908939a6)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit dcb673d23ec09af5b4cad93b34b30bc31a3495f4)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit ca3a8768ddf3766db6b4b9e261c891c7d12ecb09)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 374ecb8343f267314e77710ec46ba30e5f3817f7)
Signed-off-by: Willy Tarreau <w@1wt.eu>