BUG/MEDIUM: tcp-check: unbreak multiple connect rules again
The last connect rule used to be ignored and that was fixed by commit
248f1173f ("BUG/MEDIUM: tcp-check: single connect rule can't detect
DOWN servers") during 1.9 development. However this patch went a bit
too far by not breaking out of the loop after a pending connect(),
resulting in a series of failed connect() to be quickly skipped and
only the last one to be taken into account.

Technically speaking the series is not exactly skipped, it's just that
TCP checks suffer from a design issue which is that there is no
distinction between a new rule and this rule's completion in the
"connect" rule handling code. As such, when evaluating TCPCHK_ACT_CONNECT
a new connection is created regardless of any previous connection in
progress, and the previous result is ignored. It seems that this issue
is mostly specific to the connect action if we refer to the comments at
the top of the function, so it might be possible to durably address it
by reworking the connect state.

For now this patch does something simpler, it restores the behaviour
before the commit above consisting in breaking out of the loop when
the connection is in progress and after skipping comment rules. This
way we fall back to the default code waiting for completion.

This patch must be backported as far as 1.8 since the commit above
was backported there. Thanks to Jérôme Magnin for reporting and
bisecting this issue.

(cherry picked from commit 7df8ca6296e7f2af2cadf05830af54b998e9c196)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 7199e96c8eea193cdf158e03d37ff5a6f4b58190)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 27f14b0f273ec46a9974bcbcf4acb33b7b8d16f7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 46a8cde865a78ecc864babb1ea8a084dd90a62e8)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 9b8513d2906928282a0d46e9136a5fa92cc63283)
Signed-off-by: Willy Tarreau <w@1wt.eu>